[query/api] add try into for action
