+++ fusequery/query/src/api/rpc/flight_client.rs
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0.
 
-use std::convert::TryFrom;
+use std::convert::{TryFrom, TryInto};
 use std::sync::Arc;
 
 use anyhow::{bail, Result};
@@ -13,12 +13,11 @@ use common_datablocks::DataBlock;
 use common_datavalues::DataSchema;
 use common_planners::{Partitions, PlanNode};
 use common_streams::SendableDataBlockStream;
-use prost::Message;
 use tokio_stream::StreamExt;
+use tonic::Request;
 
 use crate::api::rpc::flight_action::DoActionAction;
 use crate::api::rpc::{DoGetAction, ExecutePlanAction, FetchPartitionAction};
-use crate::protobuf::FlightRequest;
 
 pub struct FlightClient {
     client: FlightServiceClient<tonic::transport::channel::Channel>,
@@ -30,18 +29,30 @@ impl FlightClient {
         Ok(Self { client })
     }
 
-    // Execute do_get by action.
+    /// Execute the plan in the remote action and get the block stream.
+    pub async fn execute_remote_plan_action(
+        &mut self,
+        job_id: String,
+        plan: &PlanNode,
+    ) -> Result<SendableDataBlockStream> {
+        let action = DoGetAction::ExecutePlan(ExecutePlanAction {
+            job_id: job_id.clone(),
+            plan: plan.clone(),
+        });
+        self.do_get(&action).await
+    }
+
+    // Fetch partition action and get the result.
+    pub async fn fetch_partition_action(&mut self, uuid: String, nums: u32) -> Result<Partitions> {
+        let action = DoActionAction::FetchPartition(FetchPartitionAction { uuid, nums });
+        let body = self.do_action(&action).await?;
+        let parts: Partitions = serde_json::from_slice(&body)?;
+        Ok(parts)
+    }
+
+    // Execute do_get.
     async fn do_get(&mut self, action: &DoGetAction) -> Result<SendableDataBlockStream> {
-        // Encode flight request.
-        let flight_request = FlightRequest {
-            body: serde_json::to_string(action)?,
-        };
-
-        // Encode tonic request.
-        let mut buf = vec![];
-        flight_request.encode(&mut buf)?;
-        let request = tonic::Request::new(Ticket { ticket: buf });
-
+        let request: Request<Ticket> = action.try_into()?;
         let mut stream = self.client.do_get(request).await?.into_inner();
         match stream.message().await? {
             Some(flight_data) => {
@@ -59,33 +70,9 @@ impl FlightClient {
         }
     }
 
-    /// Execute the plan in the remote and get the stream.
-    pub async fn execute_remote_plan_action(
-        &mut self,
-        job_id: String,
-        plan: &PlanNode,
-    ) -> Result<SendableDataBlockStream> {
-        let action = DoGetAction::ExecutePlan(ExecutePlanAction {
-            job_id: job_id.clone(),
-            plan: plan.clone(),
-        });
-        self.do_get(&action).await
-    }
-
-    pub async fn do_action(&mut self, action: &DoActionAction) -> Result<Vec<u8>> {
-        // Request.
-        let action_request = FlightRequest {
-            body: serde_json::to_string(&action)?,
-        };
-        let mut buf = vec![];
-        action_request.encode(&mut buf)?;
-        let tonic_request = tonic::Request::new(Action {
-            r#type: "".to_string(),
-            body: buf,
-        });
-
-        // Stream.
-        let mut stream = self.client.do_action(tonic_request).await?.into_inner();
+    async fn do_action(&mut self, action: &DoActionAction) -> Result<Vec<u8>> {
+        let request: Request<Action> = action.try_into()?;
+        let mut stream = self.client.do_action(request).await?.into_inner();
         match stream.message().await? {
             None => {
                 bail!(
@@ -95,12 +82,5 @@ impl FlightClient {
             }
             Some(resp) => Ok(resp.body),
         }
-    }
-
-    pub async fn fetch_partition_action(&mut self, uuid: String, nums: u32) -> Result<Partitions> {
-        let action = DoActionAction::FetchPartition(FetchPartitionAction { uuid, nums });
-        let body = self.do_action(&action).await?;
-        let parts: Partitions = serde_json::from_slice(&body)?;
-        Ok(parts)
     }
 }
